+++ pygrunner/tmx/__init__.py
@@ -1,0 +1,1 @@
+from pygrunner.tmx.loader import TmxLoader
